# audit/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "WB Gamification API",
        "description": "Gamification engine REST API for points, badges, levels, leaderboards, challenges, kudos, and more.",
        "version": "1.6.3",
        "contact": {
            "name": "Wbcom Designs",
            "url": "https://wbcomdesigns.com"
        }
    },
    "servers": [
        {
            "url": "http://wb-gamification.local/wp-json/wb-gamification/v1"
        }
    ],
    "paths": {
...
        "/kudos": {
            "get": {
                "summary": "Get kudos",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            },
            "post": {
                "summary": "Create kudos",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "receiver_id": {
                                        "type": "integer",
                                        "description": "User ID of the member receiving kudos. One of receiver_id or recipient_login is required."
                                    },
                                    "recipient_login": {
                                        "type": "string",
                                        "description": "Login (username) or email of the member receiving kudos. Used by frontend forms where the giver does not know the recipient user_id."
                                    },
                                    "message": {
                                        "type": "string",
                                        "description": "Optional short message (max 255 chars)."
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/kudos/{id}": {
            "delete": {
                "summary": "Delete kudos",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-WP-Nonce",
                "description": "WordPress REST API nonce for cookie-based auth."
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-WB-Gam-Key",
                "description": "API key for cross-site authentication."
            }
        }
    },
    "security": [
        {
            "cookieAuth": []
        },
        {
            "apiKeyAuth": []
        }
    ]
}